fix: dom-exception error in incognito for storejs lib (#991)
* fix: broken storejs lib in incognito

* chore: seperate util fn that uses storage module

* chore: resolve eslint error

* chore: resolve security vulnerabilities

* chore: update in storejs lib
