+++ patches/storejs+2.0.3.patch
@@ -1,0 +1,45 @@
+diff --git a/node_modules/storejs/dist/store.js b/node_modules/storejs/dist/store.js
+index b4d7933..305b930 100644
+--- a/node_modules/storejs/dist/store.js
++++ b/node_modules/storejs/dist/store.js
+@@ -42,29 +42,31 @@
+   }
+   // https://github.com/jaywcjlove/store.js/pull/8
+   // Error: QuotaExceededError
+-  function dealIncognito() {
+-    var storage = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : window.localStorage;
++  function dealIncognito(storage) {
+     var _KEY = '_Is_Incognit',
+       _VALUE = 'yes';
+     try {
++      var storage = window.localStorage;
+       storage.setItem(_KEY, _VALUE);
+       storage.removeItem(_KEY);
+     } catch (e) {
+-      Storage.prototype._data = {};
+-      Storage.prototype.setItem = function (id, val) {
++      var inMemoryStorage = {};
++      inMemoryStorage._data = {};
++      inMemoryStorage.setItem = function (id, val) {
+         return this._data[id] = String(val);
+       };
+-      Storage.prototype.getItem = function (id) {
++      inMemoryStorage.getItem = function (id) {
+         return this._data.hasOwnProperty(id) ? this._data[id] : undefined;
+       };
+-      Storage.prototype.removeItem = function (id) {
++      inMemoryStorage.removeItem = function (id) {
+         return delete this._data[id];
+       };
+-      Storage.prototype.clear = function () {
++      inMemoryStorage.clear = function () {
+         return this._data = {};
+       };
+-      storage = Storage;
+-    } finally {
++      storage = inMemoryStorage;
++
++      storage.setItem(_KEY, _VALUE);
+       if (storage.getItem(_KEY) === _VALUE) storage.removeItem(_KEY);
+     }
+     return storage;
